+++ CHANGELOG.md
@@ -895,7 +895,7 @@
 # 0.99.48
 
 * Don't rethrow `ResolutionException` inside another `ResolutionException` in case of configuration errors - the stack trace reveals the call stack so nicely, there's no need to have 10 layers of exceptions too
-
+* Marked topic-based routing configuration API as obsolete (type-based routing is perfectly capable of using raw topics if you like)
 
 
 [AndreaCuneo]: https://github.com/AndreaCuneo