+++ CHANGELOG.md
@@ -896,6 +896,7 @@
 
 * Don't rethrow `ResolutionException` inside another `ResolutionException` in case of configuration errors - the stack trace reveals the call stack so nicely, there's no need to have 10 layers of exceptions too
 * Marked topic-based routing configuration API as obsolete (type-based routing is perfectly capable of using raw topics if you like)
+* Added fallback capability to the router (i.e. the ability to configure an endpoint that is used when none of the mappings match)
 
 
 [AndreaCuneo]: https://github.com/AndreaCuneo